[add] axios refresh interceptors 수정

+++ src/api/api.ts
@@ -15,23 +15,15 @@ api.interceptors.response.use(
     const { config, response } = error;
 
     // 토큰 만료
-    if (response.status === 401) {
-      if (
-        response.data.status === 'UNAUTHORIZED' &&
-        response.data.message === 'accessToken이 만료되었습니다.'
-      ) {
-        api
-          .post('/v1/api/auth/reissue')
-          .then((res) => {
-            localStorage.setItem('token', res.data.accessToken);
-            // 기존 request 처리
-            config.headers.Authorization = `Bearer ${res.data.accessToken}`;
-            return axios(config);
-          })
-          .catch(() => {
-            localStorage.clear();
-            window.location.replace('/');
-          });
+    if (response.status === 401 && response.data.message === 'accessToken이 만료되었습니다.') {
+      try {
+        const response = await api.post('v1/api/auth/reissue');
+        config.headers.Authorization = `Beraer ${response.data.accessToken}`;
+        localStorage.setItem('tokne', response.data.accessToken);
+        return api(config);
+      } catch (err) {
+        localStorage.clear();
+        window.location.replace('/');
       }
     }
     return Promise.reject(error);
